move classes to file
